+++ apps/web/src/app/auth/LoginButton.tsx
@@ -6,7 +6,16 @@ function LoginButton({ token }: { token: string }) {
   console.log('token: ', token);
   return (
     <div>
-      <button onClick={() => signIn('credentials', { redirect: false, token: token })}>LogIn</button>
+      <button
+        onClick={() =>
+          signIn('credentials', {
+            token: token,
+            callbackUrl: 'http://localhost:3000/mypage',
+          })
+        }
+      >
+        LogIn
+      </button>
       {/* <button className="rounded-xl border bg-red-300 px-12 py-4" onClick={() => signOut()}>
         Log Out
       </button> */}
