+++ apps/web/src/auth.ts
@@ -1,0 +1,59 @@
+import type { GetServerSidePropsContext, NextApiRequest, NextApiResponse } from 'next';
+import type { NextAuthOptions } from 'next-auth';
+import { getServerSession } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import axios from 'axios';
+
+// You'll need to import and pass this
+// to `NextAuth` in `app/api/auth/[...nextauth]/route.ts`
+export const config = {
+  providers: [
+    CredentialsProvider({
+      //   id: 'github-oauth',
+
+      // The name to display on the sign in form (e.g. 'Sign in with...')
+      name: 'Credentials',
+      // The credentials is used to generate a suitable form on the sign in page.
+      // You can specify whatever fields you are expecting to be submitted.
+      // e.g. domain, username, password, 2FA token, etc.
+      // You can pass any HTML attribute to the <input> tag through the object.
+      credentials: {
+        token: { type: 'string' },
+      },
+      async authorize(credentials) {
+        if (!credentials || !credentials.token) {
+          throw new Error('No credentials provided');
+        }
+
+        try {
+          const res = await axios.get('https://api.gitanimals.org/users', {
+            headers: {
+              Authorization: `Bearer ${credentials.token}`,
+            },
+          });
+
+          return res.data;
+        } catch (error) {
+          console.error(error);
+          return null;
+        }
+      },
+    }),
+  ],
+  callbacks: {
+    async session({ session, token, user }) {
+      console.log('user: ', user);
+      return session;
+    },
+    async jwt({ token, user, account, profile, isNewUser }) {
+      return token;
+    },
+  },
+} satisfies NextAuthOptions;
+
+// Use it in server contexts
+export function auth(
+  ...args: [GetServerSidePropsContext['req'], GetServerSidePropsContext['res']] | [NextApiRequest, NextApiResponse] | []
+) {
+  return getServerSession(...args, config);
+}
